Create brainwave_signal_collection_and_transmission.go
This file defines a BrainwaveSignal struct that contains two fields: AttentionIntensity and RawData. It also defines a BrainwaveSignalCollector struct that has a DeviceID field and a Signal field of type BrainwaveSignal. The BrainwaveSignalCollector struct has two methods: CollectSignal and TransmitSignal.

The CollectSignal method simulates the process of collecting a brainwave signal from a device and stores the resulting BrainwaveSignal in the Signal field of the BrainwaveSignalCollector. The TransmitSignal method simulates the process of transmitting a brainwave signal over a network and takes the RawData field of the BrainwaveSignal as input.

This is just an example of what the brainwave_signal_collection_and_transmission.go file could look like, and you may choose to implement these functions in a different way depending on your specific needs and preferences.